Fix all occurences of printf with %ll length modifiers but 64 bit arguments to typecast the arguments to (unsigned) long long to avoid the warnings when compiling on 64 bit architectures.
(Logical change 1.306)

+++ ntfsprogs/mkntfs.c
@@ -776,7 +776,8 @@ static void dump_non_resident_attr(ATTR_RECORD *a)
 	int i;
 
 	l = sle64_to_cpu(a->lowest_vcn);
-	printf("Lowest VCN = %lli (0x%llx)\n", l, l);
+	printf("Lowest VCN = %lli (0x%llx)\n", (long long)l,
+			(unsigned long long)l);
 	l = sle64_to_cpu(a->highest_vcn);
 	printf("Highest VCN = %lli (0x%llx)\n", l, l);
 	printf("Mapping pairs array offset = 0x%x\n",
@@ -787,14 +788,18 @@ static void dump_non_resident_attr(ATTR_RECORD *a)
 		printf("Attribute is not the first extent. The following "
 				"sizes are meaningless:\n");
 	l = sle64_to_cpu(a->allocated_size);
-	printf("Allocated size = %lli (0x%llx)\n", l, l);
+	printf("Allocated size = %lli (0x%llx)\n", (long long)l,
+			(unsigned long long)l);
 	l = sle64_to_cpu(a->data_size);
-	printf("Data size = %lli (0x%llx)\n", l, l);
+	printf("Data size = %lli (0x%llx)\n", (long long)l,
+			(unsigned long long)l);
 	l = sle64_to_cpu(a->initialized_size);
-	printf("Initialized size = %lli (0x%llx)\n", l, l);
+	printf("Initialized size = %lli (0x%llx)\n",
+			(long long)l, (unsigned long long)l);
 	if (a->flags & ATTR_COMPRESSION_MASK) {
 		l = sle64_to_cpu(a->compressed_size);
-		printf("Compressed size = %lli (0x%llx)\n", l, l);
+		printf("Compressed size = %lli (0x%llx)\n",
+				(long long)l, (unsigned long long)l);
 	}
 	i = le16_to_cpu(a->mapping_pairs_offset);
 	dump_mapping_pairs_array((char*)a + i, le32_to_cpu(a->length) - i);
@@ -912,7 +917,8 @@ static void dump_mft_record(MFT_RECORD *m)
 	u = le16_to_cpu(m->usa_ofs);
 	printf("Update sequence array offset = %u (0x%x)\n", u, u);
 	printf("Update sequence array size = %u\n", le16_to_cpu(m->usa_count));
-	printf("$LogFile sequence number (lsn) = %llu\n", le64_to_cpu(m->lsn));
+	printf("$LogFile sequence number (lsn) = %llu\n",
+			(unsigned long long)le64_to_cpu(m->lsn));
 	printf("Sequence number = %u\n", le16_to_cpu(m->sequence_number));
 	printf("Reference (hard link) count = %u\n",
 						le16_to_cpu(m->link_count));
@@ -932,7 +938,8 @@ static void dump_mft_record(MFT_RECORD *m)
 	printf("Bytes allocated = %u (0x%x)\n", u, u);
 	r = le64_to_cpu(m->base_mft_record);
 	printf("Base mft record reference:\n\tMft record number = %llu\n\t"
-			"Sequence number = %u\n", MREF(r), MSEQNO(r));
+			"Sequence number = %u\n", (unsigned long long)MREF(r),
+			MSEQNO(r));
 	printf("Next attribute instance = %u\n",
 			le16_to_cpu(m->next_attr_instance));
 	a = (ATTR_RECORD*)((char*)m + le16_to_cpu(m->attrs_offset));
